Simplify the QueryStringSigner API
Instead of accepting a bucket/key pair and optional headers,
we should be opinionated. This class's primary use case is for
signing a known URL, so let's optimise for that. We can assume
that the caller's user-agent won't be sending a bunch of custom
headers, so we don't accept that hash any more either.